flaskapp/requirements.txt
alembic==1.15.2
asgiref==3.7.2
Babel==2.10.3
babelfish==0.6.0
bcc==0.29.1
bcrypt==3.2.2
beautifulsoup4==4.12.3
blinker==1.9.0
Brlapi==0.8.5
certifi==2023.11.17
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.6
colorama==0.4.6
command-not-found==0.3
cryptography==42.0.8
cssselect==1.2.0
cupshelpers==1.0
dbus-python==1.3.2
defer==1.0.6
distlib==0.3.8
distro==1.9.0
distro-info==1.7+build1
dnspython==2.2.1
dominate==2.7.0
duplicity==2.1.4
email-validator==1.2.1
Faker==14.2.0
fasteners==0.18
filelock==3.13.1
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
flask-cors==5.0.1
Flask-Migrate==4.1.0
flask-nav==0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
future==0.18.2
greenlet==1.1.3
html5lib==1.1
httplib2==0.20.4
idna==3.6
importlib-metadata==4.12.0
itsdangerous==2.2.0
Jinja2==3.1.2
language-selector==0.1
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
lxml==5.2.1
Mako==1.3.2.dev0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
monotonic==1.6
more-itertools==10.2.0
mqtt==0.0.1
netaddr==0.8.0
numpy==1.21.5
oauthlib==3.2.2
olefile==0.46
paho-mqtt==2.1.0
paramiko==2.12.0
passlib==1.7.4
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
ptyprocess==0.7.0
pyasyncore==1.0.2
pycairo==1.25.1
pycparser==2.22
pycups==2.0.1
Pygments==2.17.2
PyGObject==3.48.2
pyinotify==0.9.6
PyJWT==2.7.0
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.1.1
python-apt==2.7.7+ubuntu4
python-dateutil==2.8.2
python-debian==0.1.49+ubuntu2
python-dotenv==1.1.0
pytz==2024.1
pyxdg==0.28
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scour==0.38.2
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
SQLAlchemy==1.4.41
systemd-python==235
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.2
unattended-upgrades==0.1
urllib3==2.0.7
usb-creator==0.3.16
virtualenv==20.25.0+ds
visitor==0.1.3
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==2.0.3
WTForms==3.0.1
xdg==5
xkit==0.0.0
zipp==1.0.0
